Separate descriptor extensions by files
